created server.js file -composes middleware and routers -index.js handles connection
