+++ knexfile.js
@@ -1,0 +1,28 @@
+module.exports = {
+  development: {
+    client: 'pg',
+    connection: {
+      database: 'clothing-morgue',
+      user: 'postgres',
+      password: 'HorseC@sio82'
+    },
+    migrations: {
+      tableName: 'knex_migrations',
+      directory: './data/migrations'
+    },
+    seeds: {
+      directory: './data/seeds'
+    }
+  },
+  production: {
+    client: 'pg',
+    connection: process.env.DATABASE_URL,
+    migrations: {
+      tableName: 'knex_migrations',
+      directory: './data/migrations'
+    },
+    seeds: {
+      directory: './data/seeds'
+    }
+  }
+};
