+++ package.json
@@ -9,9 +9,14 @@
     "@babel/node": "^7.6.1",
     "@babel/preset-env": "^7.6.0",
     "@types/dotenv": "^6.1.1",
+    "cloudinary": "^1.15.0",
     "cors": "^2.8.5",
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
+    "express-form-data": "^2.0.10",
+    "knex": "^0.19.3",
+    "knex-cleaner": "^1.3.0",
+    "pg": "^7.12.1",
     "uuid": "^3.3.3"
   },
   "description": "",
