+++ src/index.js
@@ -1,33 +1,6 @@
-import "dotenv/config";
-import cors from "cors";
-import express from "express";
-import models from "./models";
-import routes from "./routes";
+import 'dotenv/config';
+import server from './server';
 
-console.log("Hello, Visual Studio programmer! \n \n Remember to have fun! \n");
+const port = process.env.PORT || 3000;
 
-console.log(process.env.MY_SECRET);
-
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use((req, res, next) => {
-  req.context = {
-    models,
-    me: models.users[1],
-  };
-  next();
-});
-app.use("/session", routes.session);
-app.use("/users", routes.users);
-app.use("/messages", routes.messages);
-
-app.get('/', (req, res) => {
-  return res.send('What are you doing here?');
-});
-
-app.listen(process.env.PORT, () =>
-  console.log(`\n Clothing Morgue listening on port ${process.env.PORT}! \n`)
-);
+server.listen(port, () => console.log(`\n Clothing Morgue listening on port ${process.env.PORT}! \n`));
